# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Tasks API",
    "description": "App para gerenciar tarfas",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [],
  "schemes": [
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
    "/api-docs": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/foo": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string",
            "description": "Token JWT"
          }
        ],
        "responses": {}
      }
    },
    "/login": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/tasks/": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "name",
            "in": "body",
            "type": "string",
            "description": "Name"
          },
          {
            "name": "user",
            "in": "body",
            "type": "string",
            "description": "user"
          }
        ],
        "responses": {}
      }
    },
    "/tasks/{id}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "body",
            "type": "string",
            "description": "Name"
          },
          {
            "name": "user",
            "in": "body",
            "type": "string",
            "description": "user"
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    }
  },
  "definitions": {
    "Task": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Programar"
        },
        "user": {
          "type": "string",
          "example": "Gordinho"
        }
      },
      "required": [
        "name",
        "user"
      ]
    }
  }
}